+++ test/support/live_views/lifecycle.ex
@@ -1,14 +1,14 @@
 defmodule Phoenix.LiveViewTest.InitAssigns do
   alias Phoenix.LiveView
 
-  def on_mount(_params, _session, socket) do
+  def on_mount(:default, _params, _session, socket) do
     {:cont,
      socket
      |> LiveView.assign(:init_assigns_mount, true)
      |> LiveView.assign(:last_on_mount, :init_assigns_mount)}
   end
 
-  def other_mount(_params, _session, socket) do
+  def on_mount(:other, _params, _session, socket) do
     {:cont,
      socket
      |> LiveView.assign(:init_assigns_other_mount, true)
@@ -19,7 +19,7 @@ end
 defmodule Phoenix.LiveViewTest.MountArgs do
   use Phoenix.Component
 
-  def on_mount(%{inlined: true}, _params, _session, socket) do
+  def on_mount(:inlined, _params, _session, socket) do
     qs = URI.encode_query(%{called: true, inlined: true})
     {:halt, push_redirect(socket, to: "/lifecycle?#{qs}")}
   end
@@ -30,7 +30,7 @@ defmodule Phoenix.LiveViewTest.HooksLive do
   alias Phoenix.LiveViewTest.InitAssigns
 
   on_mount InitAssigns
-  on_mount {InitAssigns, :other_mount}
+  on_mount {InitAssigns, :other}
 
   def render(assigns) do
     ~H"""
@@ -113,11 +113,11 @@ defmodule Phoenix.LiveViewTest.HooksLive.BadMount do
 
   on_mount {__MODULE__, :bad_mount}
 
+  def on_mount(:bad_mount, _params, _session, _socket), do: :boom
+
   def mount(_params, _session, _socket) do
     raise "expected to exit before #{__MODULE__}.mount/3"
   end
-
-  def bad_mount(_params, _session, _socket), do: :boom
 
   def render(assigns), do: ~H"<div></div>"
 end
@@ -127,7 +127,7 @@ defmodule Phoenix.LiveViewTest.HooksLive.HaltMount do
 
   on_mount {__MODULE__, :hook}
 
-  def hook(_, _, socket), do: {:halt, socket}
+  def on_mount(:hook, _, _, socket), do: {:halt, socket}
   def render(assigns), do: ~H"<div></div>"
 end
 
@@ -143,7 +143,7 @@ defmodule Phoenix.LiveViewTest.HooksLive.RedirectMount do
 
   on_mount {__MODULE__, :hook}
 
-  def hook(_, _, %{assigns: %{live_action: action}} = socket) do
+  def on_mount(:hook, _, _, %{assigns: %{live_action: action}} = socket) do
     {action, push_redirect(socket, to: "/lifecycle")}
   end
 
@@ -164,7 +164,7 @@ end
 defmodule Phoenix.LiveViewTest.HaltConnectedMount do
   alias Phoenix.LiveView
 
-  def on_mount(_params, _session, socket) do
+  def on_mount(_arg, _params, _session, socket) do
     if LiveView.connected?(socket) do
       {:halt, LiveView.push_redirect(socket, to: "/lifecycle")}
     else